# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NexusHR VX API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations/current:
    get:
      operationId: getCurrentOrganization
      tags: [organizations]
      summary: Get current organization
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    patch:
      operationId: updateOrganization
      tags: [organizations]
      summary: Update organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganization"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
...
  /roles:
    get:
      operationId: listRoles
      tags: [roles]
      summary: List AI employee roles
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: category
          in: query
          schema:
            type: string
        - name: industry
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [relevance, rating, price_asc, price_desc, newest]
      responses:
        "200":
          description: Roles list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleList"
  /roles/{id}:
    get:
      operationId: getRole
      tags: [roles]
      summary: Get role details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
...
  /interviews:
    post:
      operationId: createInterviewSession
      tags: [interviews]
      summary: Start interview session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInterview"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewSession"
  /interviews/{id}:
    get:
      operationId: getInterviewSession
      tags: [interviews]
      summary: Get interview session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewSession"
  /interviews/{id}/messages:
    post:
      operationId: sendInterviewMessage
      tags: [interviews]
      summary: Send message to interview candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendInterviewMessage"
      responses:
        "200":
          description: Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewMessageResponse"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationList"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationItem"
...
  /support/tickets:
    get:
      operationId: listTickets
      tags: [support]
      summary: List support tickets
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Tickets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketList"
    post:
      operationId: createTicket
      tags: [support]
      summary: Create support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicket"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketItem"
  /support/articles:
    get:
      operationId: listArticles
      tags: [support]
      summary: List knowledge base articles
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleList"
...
components:
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 12
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, limit, total, totalPages]
    Organization:
      type: object
      properties:
        id:
          type: integer
        clerkOrgId:
          type: string
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
        timezone:
          type: string
        createdAt:
          type: string
      required: [id, clerkOrgId, name, slug, timezone, createdAt]
    UpdateOrganization:
      type: object
      properties:
        name:
          type: string
        logoUrl:
          type: string
        industry:
          type: string
        timezone:
          type: string
...
    Role:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        department:
          type: string
        category:
          type: string
        industry:
          type: string
        seniorityLevel:
          type: string
        description:
          type: string
        coreResponsibilities:
          type: object
        tasks:
          type: object
        toolsAndIntegrations:
          type: object
          nullable: true
        performanceMetrics:
          type: object
          nullable: true
        personalityDefaults:
          type: object
          nullable: true
        skillsTags:
          type: object
          nullable: true
        priceMonthly:
          type: integer
        avatarUrl:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
      required: [id, title, department, category, industry, seniorityLevel, description, coreResponsibilities, tasks, priceMonthly]
    RoleList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Role"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [data, pagination]
...
    CreateInterview:
      type: object
      properties:
        roleId:
          type: integer
        mode:
          type: string
          enum: [text, voice]
      required: [roleId]
    InterviewCandidate:
      type: object
      properties:
        id:
          type: integer
        candidateName:
          type: string
        personalityProfile:
          type: object
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        scores:
          type: object
          nullable: true
        selected:
          type: integer
      required: [id, candidateName]
    InterviewSession:
      type: object
      properties:
        id:
          type: integer
        roleId:
          type: integer
        status:
          type: string
        mode:
          type: string
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/InterviewCandidate"
        createdAt:
          type: string
      required: [id, roleId, status, mode, candidates, createdAt]
    SendInterviewMessage:
      type: object
      properties:
        candidateId:
          type: integer
        content:
          type: string
      required: [candidateId, content]
    InterviewMessageResponse:
      type: object
      properties:
        userMessage:
          type: object
          properties:
            id:
              type: integer
            content:
              type: string
            role:
              type: string
          required: [id, content, role]
        aiMessage:
          type: object
          properties:
            id:
              type: integer
            content:
              type: string
            role:
              type: string
            audioUrl:
              type: string
              nullable: true
          required: [id, content, role]
      required: [userMessage, aiMessage]
...
    NotificationItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        title:
          type: string
        message:
          type: string
        isRead:
          type: integer
        createdAt:
          type: string
      required: [id, type, title, message, isRead, createdAt]
    NotificationList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/NotificationItem"
        pagination:
          $ref: "#/components/schemas/Pagination"
        unreadCount:
          type: integer
      required: [data, pagination, unreadCount]
    TicketItem:
      type: object
      properties:
        id:
          type: integer
        subject:
          type: string
        description:
          type: string
        category:
          type: string
          nullable: true
        priority:
          type: string
          nullable: true
        status:
          type: string
        createdAt:
          type: string
      required: [id, subject, description, status, createdAt]
    TicketList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/TicketItem"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [data, pagination]
    CreateTicket:
      type: object
      properties:
        subject:
          type: string
        description:
          type: string
        category:
          type: string
        priority:
          type: string
      required: [subject, description]
    ArticleItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        category:
          type: string
          nullable: true
        tags:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, title, content, createdAt]
    ArticleList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ArticleItem"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [data, pagination]
...